+++ area120/CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## [0.8.4](https://github.com/googleapis/google-cloud-go/compare/area120/v0.8.3...area120/v0.8.4) (2023-11-01)
+
+
+### Bug Fixes
+
+* **area120:** Bump google.golang.org/api to v0.149.0 ([8d2ab9f](https://github.com/googleapis/google-cloud-go/commit/8d2ab9f320a86c1c0fab90513fc05861561d0880))
+
 ## [0.8.3](https://github.com/googleapis/google-cloud-go/compare/area120/v0.8.2...area120/v0.8.3) (2023-10-26)