+++ asset/CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## [1.15.3](https://github.com/googleapis/google-cloud-go/compare/asset/v1.15.2...asset/v1.15.3) (2023-11-01)
+
+
+### Bug Fixes
+
+* **asset:** Bump google.golang.org/api to v0.149.0 ([8d2ab9f](https://github.com/googleapis/google-cloud-go/commit/8d2ab9f320a86c1c0fab90513fc05861561d0880))
+
 ## [1.15.2](https://github.com/googleapis/google-cloud-go/compare/asset/v1.15.1...asset/v1.15.2) (2023-10-26)